requirements.txt
torch 
torchvision 
torchaudio
transformers 
datasets 
accelerate 
pandas 
numpy 
scikit-learn
python-box
ensure
joblib
ipykernel
tqdm
pandas
numpy
streamlit
#-e .